api_yamdb/requirements.txt
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==2.2.16
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
django-filter==21.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
